removed model_refiner import from dummy data
